UPDATE
update role and database

+++ list-beras.php
@@ -112,10 +112,18 @@ $CountBeras = $row['cnt_beras'];
 										<h6><?= $CountBeras ?> Karung</h6>
 									</div>
 								</div>
-                                <hr>
-                                <div class="d-grid gap-2">
-                                    <a href="input-master-beras.php" type="button" class="btn btn-primary"><i class="bi bi-plus-circle-fill"></i> Tambah Stok Beras</a>
-                                </div>
+                                <?php
+                                
+                                if ($_SESSION['role'] == "admin") {
+                                    ?>
+                                    <hr>
+                                    <div class="d-grid gap-2">
+                                        <a href="input-master-beras.php" type="button" class="btn btn-primary"><i class="bi bi-plus-circle-fill"></i> Tambah Stok Beras</a>
+                                    </div>
+                                    <?php
+                                }
+                                
+                                ?>
 							</div>
 						</div>
 					</div>
@@ -138,7 +146,15 @@ $CountBeras = $row['cnt_beras'];
                                             <th scope="col">JUMLAH KARUNG</th>
                                             <th scope="col">BERAT (Kg)</th>
                                             <th scope="col">HARGA (/Kg)</th>
-                                            <th scope="col">ACTION</th>
+                                            <?php
+                                            
+                                            if ($_SESSION['role'] == "admin") {
+                                                ?>
+                                                <th scope="col">ACTION</th>
+                                                <?php
+                                            }
+                                            
+                                            ?>
                                         </tr>
                                     </thead>
                                     <tbody>
@@ -171,10 +187,18 @@ $CountBeras = $row['cnt_beras'];
                                                     <td style="color:<?= $ColorText ?>"><?= $KarungBeras ?></td>
                                                     <td style="font-weight:bold"><?= $BeratBeras ?></td>
                                                     <td>Rp <?= number_format($Harga) ?></td>
-                                                    <td colspan=2>
-                                                        <a href="update-master-beras.php?id=<?= $Id ?>"><span class="badge rounded-pill text-bg-info">Edit Data</span></a>
-                                                        <a href="proses.php?function=delete-master-beras&id=<?= $Id ?>" onclick="return checkDelete()"><span class="badge rounded-pill text-bg-danger">Delete Data</span></a>
-                                                    </td>
+                                                    <?php
+                                                    
+                                                    if ($_SESSION['role'] == 'admin') {
+                                                        ?>
+                                                        <td colspan=2>
+                                                            <a href="update-master-beras.php?id=<?= $Id ?>"><span class="badge rounded-pill text-bg-info">Edit Data</span></a>
+                                                            <a href="proses.php?function=delete-master-beras&id=<?= $Id ?>" onclick="return checkDelete()"><span class="badge rounded-pill text-bg-danger">Delete Data</span></a>
+                                                        </td>
+                                                        <?php
+                                                    }
+
+                                                    ?>
                                                 </tr>
             
                                                 <?php
