+++ list-padi.php
@@ -112,10 +112,18 @@ $CountPadi = $row['cnt_padi'];
 										<h6><?= $CountPadi ?> Karung</h6>
 									</div>
 								</div>
-                                <hr>
-                                <div class="d-grid gap-2">
-                                    <a href="input-master-padi.php" type="button" class="btn btn-primary"><i class="bi bi-plus-circle-fill"></i> Tambah Stok Padi</a>
-                                </div>
+                                <?php
+                                
+                                if ($_SESSION['role'] == "admin") {
+                                    ?>
+                                    <hr>
+                                    <div class="d-grid gap-2">
+                                        <a href="input-master-padi.php" type="button" class="btn btn-primary"><i class="bi bi-plus-circle-fill"></i> Tambah Stok Padi</a>
+                                    </div>
+                                    <?php
+                                }
+                                
+                                ?>
 							</div>
 						</div>
 					</div>
@@ -136,7 +144,15 @@ $CountPadi = $row['cnt_padi'];
                                             <th scope="col">JENIS PADI</th>
                                             <th scope="col">JUMLAH KARUNG</th>
                                             <th scope="col">BERAT (Kg)</th>
-                                            <th scope="col">ACTION</th>
+                                            <?php
+                                            
+                                            if ($_SESSION['role'] == "admin") {
+                                                ?>
+                                                 <th scope="col">ACTION</th>
+                                                <?php
+                                            }
+                                            
+                                            ?>
                                         </tr>
                                     </thead>
                                     <tbody>
@@ -184,10 +200,18 @@ $CountPadi = $row['cnt_padi'];
                                                     <td style="font-weight:bold"><?= $JenisPadi . " " . $KeteranganHabis ?></td>
                                                     <td style="color:<?= $ColorText ?>"><?= $KarungPadi ?></td>
                                                     <td style="font-weight:bold"><?= $BeratPadi ?></td>
-                                                    <td colspan=2>
-                                                        <a href="update-master-padi.php?id=<?= $Id ?>"><span class="badge rounded-pill text-bg-info">Edit Data</span></a>
-                                                        <a href="proses.php?function=delete-master-padi&id=<?= $Id ?>" onclick="return checkDelete()"><span class="badge rounded-pill text-bg-danger">Delete Data</span></a>
-                                                    </td>
+                                                    <?php
+                                                    
+                                                    if ($_SESSION['role'] == "admin") {
+                                                        ?>
+                                                        <td colspan=2>
+                                                            <a href="update-master-padi.php?id=<?= $Id ?>"><span class="badge rounded-pill text-bg-info">Edit Data</span></a>
+                                                            <a href="proses.php?function=delete-master-padi&id=<?= $Id ?>" onclick="return checkDelete()"><span class="badge rounded-pill text-bg-danger">Delete Data</span></a>
+                                                        </td>
+                                                        <?php
+                                                    }
+                                                    
+                                                    ?>
                                                 </tr>
             
                                                 <?php
